test: consolidate Scenario fixtures
